moved the Stopwatch classes to their own component
