+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -23,6 +23,7 @@
         "symfony/http-kernel": "2.2.*",
         "symfony/filesystem": "2.2.*",
         "symfony/routing": "2.2.*",
+        "symfony/stopwatch": "2.2.*",
         "symfony/templating": "2.2.*",
         "symfony/translation": "2.2.*",
         "doctrine/common": ">=2.2,<2.4-dev"
