+++ src/Symfony/Bundle/TwigBundle/composer.json
@@ -19,6 +19,9 @@
         "php": ">=5.3.3",
         "symfony/twig-bridge": "2.2.*"
     },
+    "require-dev": {
+        "symfony/stopwatch": "2.2.*"
+    },
     "autoload": {
         "psr-0": { "Symfony\\Bundle\\TwigBundle\\": "" }
     },
