+++ src/Symfony/Component/Stopwatch/.gitattributes
@@ -1,0 +1,2 @@
+Tests/ export-ignore
+phpunit.xml.dist export-ignore
