+++ src/Symfony/Component/Stopwatch/Stopwatch.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\HttpKernel\Debug;
+namespace Symfony\Component\Stopwatch;
 
 /**
  * Stopwatch provides a way to profile code.
