+++ src/Symfony/Component/Stopwatch/StopwatchEvent.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\HttpKernel\Debug;
+namespace Symfony\Component\Stopwatch;
 
 /**
  * Represents an Event managed by Stopwatch.
