+++ src/Symfony/Component/Stopwatch/StopwatchPeriod.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\HttpKernel\Debug;
+namespace Symfony\Component\Stopwatch;
 
 /**
  * Represents an Period for an Event.
