+++ src/Symfony/Component/Stopwatch/Tests/StopwatchEventTest.php
@@ -9,9 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\HttpKernel\Tests\Debug;
+namespace Symfony\Component\Stopwatch\Tests;
 
-use Symfony\Component\HttpKernel\Debug\StopwatchEvent;
+use Symfony\Component\Stopwatch\StopwatchEvent;
 
 /**
  * StopwatchEventTest
