+++ src/Symfony/Component/Stopwatch/composer.json
@@ -1,7 +1,7 @@
 {
-    "name": "symfony/propel1-bridge",
-    "type": "symfony-bridge",
-    "description": "Symfony Propel1 Bridge",
+    "name": "symfony/stopwatch",
+    "type": "library",
+    "description": "Symfony Stopwatch Component",
     "keywords": [],
     "homepage": "http://symfony.com",
     "license": "MIT",
@@ -16,19 +16,12 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
-        "symfony/http-foundation": "2.2.*",
-        "symfony/http-kernel": "2.2.*",
-        "symfony/form": "2.2.*",
-        "propel/propel1": "1.6.*"
-    },
-    "require-dev": {
-        "symfony/stopwatch": "2.2.*"
+        "php": ">=5.3.3"
     },
     "autoload": {
-        "psr-0": { "Symfony\\Bridge\\Propel1\\": "" }
+        "psr-0": { "Symfony\\Component\\Stopwatch\\": "" }
     },
-    "target-dir": "Symfony/Bridge/Propel1",
+    "target-dir": "Symfony/Component/Stopwatch",
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
